Remove patch and add NFT ID Storage increment to phala_world pallet
