--- rmrk-substrate/rmrk-core/README.md
@@ -1,5 +1,0 @@
-# RMRK Core Pallet
-
-Documentation [https://rmrk-team.github.io/rmrk-substrate/#/pallets/rmrk-core](https://rmrk-team.github.io/rmrk-substrate/#/pallets/rmrk-core)
-
-License: Apache-2.0
