--- rmrk-substrate/rmrk-core/src/mock.rs
@@ -1,188 +1,0 @@
-// Copyright (C) 2021-2022 RMRK
-// This file is part of rmrk-core.
-// License: Apache 2.0 modified by RMRK, see LICENSE.md
-
-use super::*;
-use crate as pallet_rmrk_core;
-
-use frame_support::{
-	parameter_types,
-	traits::{AsEnsureOriginWithArg, ConstU32, Everything, Locker},
-	weights::Weight,
-};
-use frame_system::EnsureRoot;
-use sp_core::{crypto::AccountId32, H256};
-use sp_runtime::{
-	testing::Header,
-	traits::{BlakeTwo256, IdentityLookup},
-	Perbill,
-};
-
-mod nfc {
-	// Re-export needed for `impl_outer_event!`.
-	pub use super::super::*;
-}
-
-type AccountId = AccountId32;
-type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
-type Block = frame_system::mocking::MockBlock<Test>;
-type Balance = u128;
-
-// Configure a mock runtime to test the pallet.
-frame_support::construct_runtime!(
-	pub enum Test where
-		Block = Block,
-		NodeBlock = Block,
-		UncheckedExtrinsic = UncheckedExtrinsic,
-	{
-		System: frame_system::{Pallet, Call, Config, Storage, Event<T>},
-		Uniques: pallet_uniques::{Pallet, Storage, Event<T>},
-		RmrkCore: pallet_rmrk_core::{Pallet, Call, Event<T>},
-		Balances: pallet_balances::{Pallet, Call, Storage, Config<T>, Event<T>},
-	}
-);
-
-parameter_types! {
-	pub ClassBondAmount: Balance = 100;
-	pub MaxMetadataLength: u32 = 256;
-	pub const MaxRecursions: u32 = 4;
-	pub const ResourceSymbolLimit: u32 = 10;
-	pub const PartsLimit: u32 = 50;
-	pub const MaxPriorities: u32 = 3;
-	pub const CollectionSymbolLimit: u32 = 100;
-	pub const MaxResourcesOnMint: u32 = 3;
-}
-
-impl pallet_rmrk_core::Config for Test {
-	// type Currency = Balances;
-	type Event = Event;
-	type ProtocolOrigin = EnsureRoot<AccountId>;
-	type MaxRecursions = MaxRecursions;
-	type ResourceSymbolLimit = ResourceSymbolLimit;
-	type PartsLimit = PartsLimit;
-	type MaxPriorities = MaxPriorities;
-	type CollectionSymbolLimit = CollectionSymbolLimit;
-	type MaxResourcesOnMint = MaxResourcesOnMint;
-}
-
-parameter_types! {
-	pub const CollectionDeposit: Balance = 10_000 * RMRK; // 1 UNIT deposit to create asset class
-	pub const ItemDeposit: Balance = 100 * RMRK; // 1/100 UNIT deposit to create asset instance
-	pub const KeyLimit: u32 = 32;	// Max 32 bytes per key
-	pub const ValueLimit: u32 = 64;	// Max 64 bytes per value
-	pub const UniquesMetadataDepositBase: Balance = 1000 * RMRK;
-	pub const AttributeDepositBase: Balance = 100 * RMRK;
-	pub const DepositPerByte: Balance = 10 * RMRK;
-	pub const UniquesStringLimit: u32 = 32;
-}
-
-impl pallet_uniques::Config for Test {
-	type Event = Event;
-	type CollectionId = u32;
-	type ItemId = u32;
-	type Currency = Balances;
-	type ForceOrigin = EnsureRoot<AccountId>;
-	type CreateOrigin = AsEnsureOriginWithArg<frame_system::EnsureSigned<AccountId>>;
-	type Locker = pallet_rmrk_core::Pallet<Test>;
-	type CollectionDeposit = CollectionDeposit;
-	type ItemDeposit = ItemDeposit;
-	type MetadataDepositBase = UniquesMetadataDepositBase;
-	type AttributeDepositBase = AttributeDepositBase;
-	type DepositPerByte = DepositPerByte;
-	type StringLimit = UniquesStringLimit;
-	type KeyLimit = KeyLimit;
-	type ValueLimit = ValueLimit;
-	type WeightInfo = ();
-	// type InstanceReserveStrategy = NFT;
-}
-
-parameter_types! {
-	pub const BlockHashCount: u64 = 250;
-	pub const MaximumBlockWeight: Weight = 1024;
-	pub const MaximumBlockLength: u32 = 2 * 1024;
-	pub const AvailableBlockRatio: Perbill = Perbill::one();
-}
-
-impl frame_system::Config for Test {
-	type BaseCallFilter = Everything;
-	type BlockWeights = ();
-	type BlockLength = ();
-	type DbWeight = ();
-	type Origin = Origin;
-	type Call = Call;
-	type Index = u64;
-	type BlockNumber = u64;
-	type Hash = H256;
-	type Hashing = BlakeTwo256;
-	type AccountId = AccountId;
-	type Lookup = IdentityLookup<Self::AccountId>;
-	type Header = Header;
-	type Event = Event;
-	type BlockHashCount = BlockHashCount;
-	type Version = ();
-	type PalletInfo = PalletInfo;
-	type AccountData = pallet_balances::AccountData<Balance>;
-	type OnNewAccount = ();
-	type OnKilledAccount = ();
-	type SystemWeightInfo = ();
-	type SS58Prefix = ();
-	type OnSetCode = ();
-	type MaxConsumers = ConstU32<2>;
-}
-
-parameter_types! {
-	pub const ExistentialDeposit: u64 = 1;
-	pub const MaxReserves: u32 = 50;
-}
-impl pallet_balances::Config for Test {
-	type Balance = Balance;
-	type Event = Event;
-	type DustRemoval = ();
-	type ExistentialDeposit = ExistentialDeposit;
-	type AccountStore = frame_system::Pallet<Test>;
-	type MaxLocks = ();
-	type WeightInfo = ();
-	type MaxReserves = MaxReserves;
-	// type ReserveIdentifier = ReserveIdentifier;
-	type ReserveIdentifier = ();
-}
-
-pub const ALICE: AccountId = AccountId::new([1u8; 32]);
-pub const BOB: AccountId = AccountId::new([2u8; 32]);
-pub const CHARLIE: AccountId = AccountId::new([3u8; 32]);
-pub const RMRK: Balance = 1;
-pub const COLLECTION_ID_0: <Test as pallet_uniques::Config>::CollectionId = 0;
-// pub const COLLECTION_ID_1: <Test as pallet_uniques::Config>::CollectionId = 1;
-pub const NFT_ID_0: <Test as pallet_uniques::Config>::ItemId = 0;
-pub const NOT_EXISTING_CLASS_ID: <Test as pallet_uniques::Config>::CollectionId = 999;
-pub const RESOURCE_ZERO: ResourceId = 0;
-pub const RESOURCE_ONE: ResourceId = 1;
-pub const RESOURCE_TWO: ResourceId = 2;
-pub const MAX_BURNS: u32 = 4;
-
-pub struct ExtBuilder;
-impl Default for ExtBuilder {
-	fn default() -> Self {
-		ExtBuilder
-	}
-}
-
-impl ExtBuilder {
-	pub fn build(self) -> sp_io::TestExternalities {
-		let mut t = frame_system::GenesisConfig::default().build_storage::<Test>().unwrap();
-
-		pallet_balances::GenesisConfig::<Test> {
-			balances: vec![
-				(ALICE, 20_000_000 * RMRK),
-				(BOB, 15_000 * RMRK),
-				(CHARLIE, 150_000 * RMRK),
-			],
-		}
-		.assimilate_storage(&mut t)
-		.unwrap();
-
-		let mut ext = sp_io::TestExternalities::new(t);
-		ext.execute_with(|| System::set_block_number(1));
-		ext
-	}
-}
